lab2

+++ src/main/java/miu/edu/demo/service/PostService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface PostService {
 
     List<PostDto> findAll();
-    PostDto getPostById(int id);
+    PostDto getPostById(long id);
     void save(PostDto p);
-    void delete(int id);
+    void delete(long id);
     void update(int id, PostDto p);
 
     List<PostDto> findAllPostsByAuthor(String author);
